Moved spotify api code to spotify.js
